+++ old/modules/component.js
@@ -1,3 +1,7 @@
+// ## Component
+//
+// Represents a component, view, or fragment.
+//
 Combo.Component = class {
 	// **constructor**
 	//
@@ -5,9 +9,20 @@ Combo.Component = class {
 	//
 	constructor(options = {}) {
 		//
-		// Extend the component from the options.
+		// Extend the properties from options.
 		//
 		Object.assign(this, options);
+
+		//
+		// Setup the rendered lifecycle hook.
+		//
+		if(typeof this.rendered === "function") {
+			this.render = function() {
+				options.render();
+				this.rendered();
+			};
+		}
+
 		//
 		// Invoke the created lifecycle hook.
 		//
@@ -15,10 +30,11 @@ Combo.Component = class {
 			this.created();
 		}
 	}
+
 	// **clone**
 	//
-	// Return a new instance of the component.
-	//
+	// Returns a new instance of the component.
+	//	
 	clone() {
 		var clone = Object.assign(Object.create(this), this);
 		//
@@ -27,50 +43,35 @@ Combo.Component = class {
 		if(typeof this.cloned === "function") {
 			this.cloned();
 		}
-		//
-		// Return the instance of the component.
-		//
+
 		return clone;
 	}
+
 	// **update**
 	//
-	// Update the component's data and redraw it.
+	// Update the data, then redraw the component if it's mounted.
 	//
 	update(values = {}) {
 		this.data = Object.assign({}, this.data, values);
+
 		//
 		// Invoke the updated lifecycle hook.
 		//
 		if(typeof this.updated === "function") {
 			this.updated();
 		}
+
 		//
 		// Redraw the component if it's mounted.
 		//
 		if(this.isMounted) {
-			this.mount(this.el);
-		}
-	}	
-	// **mount**
-	//
-	// Mount the component to a container element.
-	//
-	mount(el) {
-		this.el = el;
-		//
-		// Replace the HTML of the container element.
-		//
-		replaceHTML(el, this.render());
-		//
-		// Invoke the mounted lifecycle hook.
-		//
-		if(typeof this.mounted === "function") {
-			this.mounted();
+			Combo.mount(this.el, this);
 		}
 	}
+
 	// **isMounted**
 	//
-	// Determining if the component is mounted.
+	// Returns a boolean value determining if the component was mounted.
 	//
 	get isMounted() {
 		return !!this.el;